pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "intelligent-code-eval-openenv"
version = "0.1.0"
description = "Deterministic OpenEnv environment for intelligent code evaluation and optimization"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn==0.35.0",
  "pydantic==2.11.7",
  "PyYAML==6.0.2"
]

[project.optional-dependencies]
llm = [
  "groq>=0.11.0",
  "python-dotenv>=1.0.1"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
py-modules = ["api", "env", "models", "tasks", "validate_openenv"]
